Add files via upload
X's were Close and Force Index

The force index uses price and volume to determine a trend and the strength of the trend.   A shorter force index can be used to determine the short-term trend, while a longer force index, for example, a 100-day force index can be used to determine the long-term trend in prices.

A force index can also be used to identify corrections in a given trend. To do so, it can be used in conjunction with a trend following indicator.  For example, one can use a 22-day EMA for trend and a 2-day force index to identify corrections in the trend.

We can also use the force index to spot the breakouts.  The breakouts are usually confirmed by the volume and the force index takes both price and volume into account. Sudden spikes in the direction of the price moment can help confirm the breakout.